Move generics into core project
